+++ definitions/splunk_installer.rb
@@ -51,13 +51,14 @@ define :splunk_installer, url: nil do
   end
 
   local_package_resource = case node['platform_family']
-                           when 'rhel', 'fedora'  then :rpm_package
-                           when 'debian'          then :dpkg_package
-                           when 'omnios'          then :solaris_package
+                           when 'rhel', 'fedora', 'suse', 'amazon'  then :rpm_package
+                           when 'debian'                            then :dpkg_package
+                           when 'omnios'                            then :solaris_package
                            end
 
   declare_resource local_package_resource, params[:name] do
     source cached_package.gsub(/\.Z/, '')
+    version package_file[/#{params[:name]}-([^-]+)/, 1]
     options pkgopts.join(' ') if platform?('omnios')
   end
 end