+++ libraries/splunk_app_provider.rb
@@ -26,7 +26,7 @@ class Chef
     class SplunkApp < Chef::Provider::LWRPBase
       provides :splunk_app
 
-      use_inline_resources
+      use_inline_resources # ~FC113
 
       def whyrun_supported?
         true