+++ spec/recipes/install_server_spec.rb
@@ -6,7 +6,7 @@ describe 'chef-splunk::install_server' do
       ChefSpec::ServerRunner.new(
         step_into: ['splunk_installer'],
         platform: 'ubuntu',
-        version: '12.04'
+        version: '16.04'
       ) do |node|
         node.normal['splunk']['server']['url'] = 'http://splunk.example.com/server/package.deb'
       end.converge(described_recipe)
@@ -28,7 +28,7 @@ describe 'chef-splunk::install_server' do
       ChefSpec::ServerRunner.new(
         step_into: ['splunk_installer'],
         platform: 'centos',
-        version: '6.4'
+        version: '6.9'
       ) do |node|
         node.normal['splunk']['server']['url'] = 'http://splunk.example.com/server/package.rpm'
       end.converge(described_recipe)
@@ -50,7 +50,7 @@ describe 'chef-splunk::install_server' do
       ChefSpec::ServerRunner.new(
         step_into: ['splunk_installer'],
         platform: 'omnios',
-        version: '151002'
+        version: '151018'
       ) do |node|
         node.normal['splunk']['server']['url'] = 'http://splunk.example.com/server/package.pkg.Z'
       end.converge(described_recipe)